pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-brief-agent"
version = "0.1.0"
description = "Daily AI news brief: ingest, filter, score, cluster, summarize."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AI News Brief" }]
dependencies = [
    "openai>=1.50.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "tenacity>=8.2",
    "feedparser>=6.0",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "pyyaml>=6.0",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "responses>=0.25",
]
postgres = [
    "psycopg[binary]>=3.1",
]

[project.scripts]
ai-news-brief = "news_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
news_agent = ["prompts/*.txt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
